+++ src/services/cssSelectionRange.ts
@@ -6,27 +6,34 @@
 
 import { Range, Position, TextDocument } from 'vscode-languageserver-types';
 import { Stylesheet, NodeType } from '../parser/cssNodes';
+import { SelectionRange, SelectionRangeKind } from '../cssLanguageTypes';
 
-export function getSelectionRanges(document: TextDocument, position: Position, stylesheet: Stylesheet): Range[] {
-	const applicableRanges = getApplicableRanges(document, position, stylesheet);
-	const ranges = applicableRanges.map(pair => {
-		return Range.create(
-			document.positionAt(pair[0]),
-			document.positionAt(pair[1])
-		);
-	});
-	return ranges;
+export function getSelectionRanges(document: TextDocument, positions: Position[], stylesheet: Stylesheet): SelectionRange[][] {
+	function getSelectionRange(position: Position): SelectionRange[] {
+		const applicableRanges = getApplicableRanges(document, position, stylesheet);
+		const ranges = applicableRanges.map(pair => {
+			return {
+				range: Range.create(
+					document.positionAt(pair[0]),
+					document.positionAt(pair[1])
+				),
+				kind: SelectionRangeKind.Statement
+			};
+		});
+		return ranges;
+	}
+	return positions.map(getSelectionRange);
 }
 
 export function getApplicableRanges(document: TextDocument, position: Position, stylesheet: Stylesheet): number[][] {
 	let currNode = stylesheet.findChildAtOffset(document.offsetAt(position), true);
-	
+
 	if (!currNode) {
 		return [];
 	}
-	
+
 	let result = [];
-	
+
 	while (currNode) {
 		if (
 			currNode.parent &&
@@ -40,11 +47,11 @@ export function getApplicableRanges(document: TextDocument, position: Position, 
 		if (currNode.type === NodeType.Declarations) {
 			result.push([currNode.offset + 1, currNode.end - 1]);
 		} else {
-			result.push([currNode.offset, currNode.end]); 
+			result.push([currNode.offset, currNode.end]);
 		}
 
 		currNode = currNode.parent;
 	}
-	
+
 	return result;
 }